requirements.txt
pydantic>=2.7
python-dotenv>=1.0
rich>=13.7
typer>=0.12

# LangGraph / LangChain (agent + graph)
langgraph>=0.2.30
langchain>=0.2.15
langchain-community>=0.2.12

# Optional memory backends
redis>=5.0
chromadb>=0.5.5

